Move rose app code into its own subpackage
