+++ src/CSET/workflow/files/app/fetch_fcst/bin/fetch-data-filesystem.py
@@ -2,6 +2,6 @@
 
 """Retrieve files from the filesystem."""
 
-from CSET.workflow.fetch_data import FilesystemFileRetriever, fetch_data
+from CSET.rose_apps.fetch_data import FilesystemFileRetriever, fetch_data
 
 fetch_data(FilesystemFileRetriever)
