+++ src/CSET/workflow/files/app/finish_website/bin/finish-website.py
@@ -2,6 +2,6 @@
 
 """Write finished status to website front page."""
 
-import CSET.workflow.finish_website
+import CSET.rose_apps.finish_website
 
 CSET.workflow.finish_website.run()
