+++ src/CSET/workflow/files/app/preprocess_data/bin/preprocess.py
@@ -2,6 +2,6 @@
 
 """Preprocess forecast data into a single file per model."""
 
-import CSET.workflow.preprocess
+import CSET.rose_apps.preprocess
 
 CSET.workflow.preprocess.run()
