+++ tests/rose_apps/test_install_workflow.py
